refactor: migrate cliChecksumTable to CSS modules
